Move file based persistence into seperate file
